LayoutTests/media/media-source/media-source-append-acb-no-frame-lost-expected.txt


EXPECTED (source.readyState == 'closed') OK
EVENT(sourceopen)
RUN(sourceBuffer = source.addSourceBuffer("video/mock; codecs=mock"))
RUN(sourceBuffer.appendBuffer(initSegment))
EVENT(updateend)
RUN(sourceBuffer.appendBuffer(syncSampleRun(0, 10)))
EVENT(updateend)
RUN(sourceBuffer.appendBuffer(syncSampleRun(20, 30)))
EVENT(updateend)
RUN(sourceBuffer.appendBuffer(syncSampleRun(10, 20)))
EVENT(updateend)
EXPECTED (bufferedSamples.length == '30') OK
EXPECTED (enqueuedSamples.length == '30') OK
{PTS({0/1 = 0.000000}), DTS({0/1 = 0.000000}), duration({1/1 = 1.000000}), flags(1), generation(0)}
{PTS({1/1 = 1.000000}), DTS({1/1 = 1.000000}), duration({1/1 = 1.000000}), flags(0), generation(0)}
{PTS({2/1 = 2.000000}), DTS({2/1 = 2.000000}), duration({1/1 = 1.000000}), flags(0), generation(0)}
{PTS({3/1 = 3.000000}), DTS({3/1 = 3.000000}), duration({1/1 = 1.000000}), flags(0), generation(0)}
{PTS({4/1 = 4.000000}), DTS({4/1 = 4.000000}), duration({1/1 = 1.000000}), flags(0), generation(0)}
{PTS({5/1 = 5.000000}), DTS({5/1 = 5.000000}), duration({1/1 = 1.000000}), flags(0), generation(0)}
{PTS({6/1 = 6.000000}), DTS({6/1 = 6.000000}), duration({1/1 = 1.000000}), flags(0), generation(0)}
{PTS({7/1 = 7.000000}), DTS({7/1 = 7.000000}), duration({1/1 = 1.000000}), flags(0), generation(0)}
{PTS({8/1 = 8.000000}), DTS({8/1 = 8.000000}), duration({1/1 = 1.000000}), flags(0), generation(0)}
{PTS({9/1 = 9.000000}), DTS({9/1 = 9.000000}), duration({1/1 = 1.000000}), flags(0), generation(0)}
{PTS({10/1 = 10.000000}), DTS({10/1 = 10.000000}), duration({1/1 = 1.000000}), flags(1), generation(0)}
{PTS({11/1 = 11.000000}), DTS({11/1 = 11.000000}), duration({1/1 = 1.000000}), flags(0), generation(0)}
{PTS({12/1 = 12.000000}), DTS({12/1 = 12.000000}), duration({1/1 = 1.000000}), flags(0), generation(0)}
{PTS({13/1 = 13.000000}), DTS({13/1 = 13.000000}), duration({1/1 = 1.000000}), flags(0), generation(0)}
{PTS({14/1 = 14.000000}), DTS({14/1 = 14.000000}), duration({1/1 = 1.000000}), flags(0), generation(0)}
{PTS({15/1 = 15.000000}), DTS({15/1 = 15.000000}), duration({1/1 = 1.000000}), flags(0), generation(0)}
{PTS({16/1 = 16.000000}), DTS({16/1 = 16.000000}), duration({1/1 = 1.000000}), flags(0), generation(0)}
{PTS({17/1 = 17.000000}), DTS({17/1 = 17.000000}), duration({1/1 = 1.000000}), flags(0), generation(0)}
{PTS({18/1 = 18.000000}), DTS({18/1 = 18.000000}), duration({1/1 = 1.000000}), flags(0), generation(0)}
{PTS({19/1 = 19.000000}), DTS({19/1 = 19.000000}), duration({1/1 = 1.000000}), flags(0), generation(0)}
{PTS({20/1 = 20.000000}), DTS({20/1 = 20.000000}), duration({1/1 = 1.000000}), flags(1), generation(0)}
{PTS({21/1 = 21.000000}), DTS({21/1 = 21.000000}), duration({1/1 = 1.000000}), flags(0), generation(0)}
{PTS({22/1 = 22.000000}), DTS({22/1 = 22.000000}), duration({1/1 = 1.000000}), flags(0), generation(0)}
{PTS({23/1 = 23.000000}), DTS({23/1 = 23.000000}), duration({1/1 = 1.000000}), flags(0), generation(0)}
{PTS({24/1 = 24.000000}), DTS({24/1 = 24.000000}), duration({1/1 = 1.000000}), flags(0), generation(0)}
{PTS({25/1 = 25.000000}), DTS({25/1 = 25.000000}), duration({1/1 = 1.000000}), flags(0), generation(0)}
{PTS({26/1 = 26.000000}), DTS({26/1 = 26.000000}), duration({1/1 = 1.000000}), flags(0), generation(0)}
{PTS({27/1 = 27.000000}), DTS({27/1 = 27.000000}), duration({1/1 = 1.000000}), flags(0), generation(0)}
{PTS({28/1 = 28.000000}), DTS({28/1 = 28.000000}), duration({1/1 = 1.000000}), flags(0), generation(0)}
{PTS({29/1 = 29.000000}), DTS({29/1 = 29.000000}), duration({1/1 = 1.000000}), flags(0), generation(0)}
END OF TEST
